+++ src/main/java/com/example/healthcare/account/service/UserSyService.java
@@ -3,8 +3,8 @@ package com.example.healthcare.account.service;
 import com.example.healthcare.account.domain.User;
 import com.example.healthcare.account.repository.UserRepository;
 import com.example.healthcare.account.service.dto.ChangeAuthorityDTO;
-import com.example.healthcare.util.AdminCheck;
-import jakarta.persistence.EntityNotFoundException;
+import com.example.healthcare.common.exception.ResourceException;
+import com.example.healthcare.common.exception.ResourceException.ResourceExceptionCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,23 +13,19 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class UserSyService {
 
-    private final AdminCheck adminCheck;
-
-    private final UserRepository userRepository;
+  private final UserRepository userRepository;
 
 
+  @Transactional
+  public void changeAuthority(Long userId, ChangeAuthorityDTO dto) {
+    User user = userRepository.findById(userId).orElseThrow(
+      () -> new ResourceException(ResourceExceptionCode.RESOURCE_NOT_FOUND));
 
-    @Transactional
-    public void changeAuthority(ChangeAuthorityDTO dto) {
-        adminCheck.adminPage();
-        Long userId = dto.userId();
-        User user = userRepository.findById(userId).orElseThrow(
-                () -> new EntityNotFoundException("User not found with id: " + userId));
-        if(user.getAuthorityType().equals(dto.authorityType())) {
-            throw new IllegalArgumentException("User authority type does not match expected authority type");
-        }
-        user.changeAuthority(dto.authorityType());
-        userRepository.save(user);
-
+    if (user.getAuthorityType().equals(dto.authorityType())) {
+      return;
     }
+
+    user.changeAuthority(dto.authorityType());
+    userRepository.save(user);
+  }
 }
