+++ src/main/java/com/example/healthcare/account/service/dto/ReissueTokenDTO.java
@@ -1,9 +1,10 @@
 package com.example.healthcare.account.service.dto;
 
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 
 public record ReissueTokenDTO(
-        @NotBlank
-        String userId,
+        @NotNull
+        String email,
         @NotBlank
         String refreshToken) {}
