+++ src/main/java/com/example/healthcare/config/security/jwt/JwtAuthorizationFilter.java
@@ -1,7 +1,7 @@
-package com.example.healthcare.security.jwt;
+package com.example.healthcare.config.security.jwt;
 
-import com.example.healthcare.security.user.UserDetailsImpl;
-import com.example.healthcare.security.user.UserDetailsServiceImpl;
+import com.example.healthcare.config.security.user.LoginUser;
+import com.example.healthcare.config.security.user.UserDetailsServiceImpl;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -17,9 +17,6 @@ import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
-
-import static com.example.healthcare.security.jwt.JwtUtil.ACCESS_TOKEN;
-import static com.example.healthcare.security.jwt.JwtUtil.BEARER_PREFIX;
 
 @Slf4j(topic = "JWT 검증 및 인가")
 @RequiredArgsConstructor
@@ -60,14 +57,14 @@ public class JwtAuthorizationFilter extends OncePerRequestFilter {
 
     // 인증 객체 생성
     private Authentication createAuthentication(String username) {
-        UserDetailsImpl userDetails = (UserDetailsImpl) userDetailsService.loadUserByUsername(username);
+        LoginUser userDetails = (LoginUser) userDetailsService.loadUserByUsername(username);
         return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
     }
 
     // header 에서 ACCESS TOKEN 가져오는 메서드
     private String getJwtFromHeader(HttpServletRequest request) {
-        String accessToken = request.getHeader(ACCESS_TOKEN);
-        if (StringUtils.hasText(accessToken) && accessToken.startsWith(BEARER_PREFIX)) {
+        String accessToken = request.getHeader(JwtUtil.ACCESS_TOKEN);
+        if (StringUtils.hasText(accessToken) && accessToken.startsWith(JwtUtil.BEARER_PREFIX)) {
             return accessToken.substring(7);
         }
         return null;
