+++ src/main/java/com/example/healthcare/config/security/user/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.healthcare.security.user;
+package com.example.healthcare.config.security.user;
 
 import com.example.healthcare.account.domain.User;
 import com.example.healthcare.account.repository.UserRepository;
@@ -19,13 +19,13 @@ public class UserDetailsServiceImpl implements UserDetailsService {
         User user = userRepository.findByEmail(email)
             .orElseThrow(() -> new UsernameNotFoundException("Not Found " + email));
 
-        return new UserDetailsImpl(user);
+        return new LoginUser(user);
     }
 
-    public UserDetailsImpl loadUserById(Long id){
+    public LoginUser loadUserById(Long id){
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UsernameNotFoundException("Can't find " + id));
 
-        return new UserDetailsImpl(user);
+        return new LoginUser(user);
     }
 }
