--- src/main/java/com/example/healthcare/security/user/UserDetailsImpl.java
@@ -1,67 +1,0 @@
-package com.example.healthcare.security.user;
-
-import com.example.healthcare.account.domain.User;
-import com.example.healthcare.account.domain.code.AuthorityType;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-public class UserDetailsImpl implements UserDetails {
-
-    private final User user;
-
-    public UserDetailsImpl(User user) {
-        this.user = user;
-    }
-
-    @JsonProperty
-    public User getUser() {
-        return user;
-    }
-    @Override
-    @JsonProperty
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    @JsonProperty
-    public String getUsername() {
-        return user.getName();
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        String authority = String.valueOf(user.getAuthorityType());
-
-        SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(authority);
-        Collection<GrantedAuthority> authorities = new ArrayList<>();
-        authorities.add(simpleGrantedAuthority);
-
-        return authorities;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
