--- src/main/resources/application.properties
@@ -1,13 +1,0 @@
-spring.application.name=health-care
-
-spring.profiles.include=security
-
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.jpa.hibernate.ddl-auto=update
-
-spring.jpa.properties.hibernate.show_sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.use_sql_comments=true
-
-spring.main.allow-bean-definition-overriding=true
